chore(*): refactor e2e test
